requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
tensorflow==2.17.0
keras==3.4.1
plotly==5.24.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
protobuf==4.25.3


